[exporter/otlphttp] split module into its own module
Moved `exporter/otlphttpexporter` to `exporters/otlphttpexporter` and making it available as its own module.

Fixes #6196
