+++ .chloggen/otlphttp-exporter-module.yaml
@@ -1,0 +1,19 @@
+# One of 'breaking', 'deprecation', 'new_component', 'enhancement', 'bug_fix'
+change_type: breaking
+
+# The name of the component, or a single word describing the area of concern, (e.g. otlpreceiver)
+component: otlphttpexporter
+
+# A brief description of the change.  Surround your text with quotes ("") if it needs to start with a backtick (`).
+note: OTLP HTTP exporter now available as its own module
+
+# One or more tracking issues or pull requests related to the change
+issues: [6196]
+
+# (Optional) One or more lines of additional information to render under the primary note.
+# These lines will be padded with 2 spaces and then inserted directly into the document.
+# Use pipe (|) for multiline entries.
+subtext: |-
+  The import path has changed from `go.opentelemetry.io/collector/exporter/otlphttpexporter` in the collector
+  module to `go.opentelemetry.io/collector/exporters/otlphttpexporter`.
+
