+++ exporters/otlphttpexporter/testdata/bad_empty_config.yaml
