+++ versions.yaml
@@ -19,6 +19,7 @@ module-sets:
       - go.opentelemetry.io/collector
       - go.opentelemetry.io/collector/cmd/builder
       - go.opentelemetry.io/collector/exporters/otlpgrpcexporter
+      - go.opentelemetry.io/collector/exporters/otlphttpexporter
       - go.opentelemetry.io/collector/semconv
       - go.opentelemetry.io/collector/pdata
